+++ Programming/Arduino/libraries/JHArduino/src/types.h
@@ -5,6 +5,6 @@
 
 const unsigned MAX_VARIABLE_NAME = 32;
 
-typedef void (*T_UPDATE_EVENT)(int oldValue, int newValue);
+typedef void (*T_UPDATE_EVENT)(long oldValue, long newValue);
 
 #endif //Types_h
